# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Minecraft Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BotStatusState =
  (typeof BotStatusState)[keyof typeof BotStatusState];

export const BotStatusState = {
  IDLE: "IDLE",
  FOLLOW: "FOLLOW",
  GUARD: "GUARD",
  COMBAT: "COMBAT",
  AUTONOMOUS: "AUTONOMOUS",
  DISCONNECTED: "DISCONNECTED",
} as const;

export type BotStatusCpuMode =
  (typeof BotStatusCpuMode)[keyof typeof BotStatusCpuMode];

export const BotStatusCpuMode = {
  LOW: "LOW",
  NORMAL: "NORMAL",
  HIGH: "HIGH",
} as const;

export interface BotStatus {
  connected: boolean;
  state: BotStatusState;
  health: number;
  food: number;
  ping: number;
  username: string;
  serverHost: string;
  serverPort: number;
  nearbyPlayers: number;
  currentTarget: string | null;
  ownerOnline: boolean;
  cpuMode: BotStatusCpuMode;
  autonomousMode: boolean;
  uptime: number;
}

export interface ConnectRequest {
  host: string;
  port: number;
  username: string;
  owner: string;
}

export type CommandRequestCommand =
  (typeof CommandRequestCommand)[keyof typeof CommandRequestCommand];

export const CommandRequestCommand = {
  follow: "follow",
  guard: "guard",
  stop: "stop",
  attack_nearest: "attack_nearest",
  toggle_autonomous: "toggle_autonomous",
  set_owner: "set_owner",
} as const;

export interface CommandRequest {
  command: CommandRequestCommand;
  value?: string | null;
}

export type BotSettingsCpuMode =
  (typeof BotSettingsCpuMode)[keyof typeof BotSettingsCpuMode];

export const BotSettingsCpuMode = {
  LOW: "LOW",
  NORMAL: "NORMAL",
  HIGH: "HIGH",
} as const;

export interface BotSettings {
  followDistance: number;
  detectionRadius: number;
  aggressionLevel: number;
  attackDelay: number;
  scanInterval: number;
  cpuMode: BotSettingsCpuMode;
  autoReconnect: boolean;
  reconnectDelay: number;
  owner: string;
}

export interface ActionResponse {
  success: boolean;
  message: string;
}

export type LogEntryLevel = (typeof LogEntryLevel)[keyof typeof LogEntryLevel];

export const LogEntryLevel = {
  info: "info",
  warn: "warn",
  error: "error",
  combat: "combat",
  state: "state",
  connection: "connection",
} as const;

export interface LogEntry {
  id: string;
  timestamp: string;
  level: LogEntryLevel;
  message: string;
}

export interface LogsResponse {
  logs: LogEntry[];
}

export interface WhitelistEntry {
  username: string;
}

export interface WhitelistResponse {
  whitelist: string[];
}

export type BotTaskType = (typeof BotTaskType)[keyof typeof BotTaskType];

export const BotTaskType = {
  follow: "follow",
  guard_area: "guard_area",
  move_to: "move_to",
} as const;

export type BotTaskStatus = (typeof BotTaskStatus)[keyof typeof BotTaskStatus];

export const BotTaskStatus = {
  pending: "pending",
  active: "active",
  paused: "paused",
  done: "done",
} as const;

export type BotTaskParams = { [key: string]: unknown };

export interface BotTask {
  id: string;
  type: BotTaskType;
  status: BotTaskStatus;
  params: BotTaskParams;
}

export type CreateTaskRequestType =
  (typeof CreateTaskRequestType)[keyof typeof CreateTaskRequestType];

export const CreateTaskRequestType = {
  follow: "follow",
  guard_area: "guard_area",
  move_to: "move_to",
} as const;

export type CreateTaskRequestParams = { [key: string]: unknown };

export interface CreateTaskRequest {
  type: CreateTaskRequestType;
  params: CreateTaskRequestParams;
}

export interface TaskQueueResponse {
  tasks: BotTask[];
}

export interface ServerConfig {
  id?: string;
  name: string;
  host: string;
  port: number;
  username: string;
  owner: string;
}

export interface ServerConfigsResponse {
  configs: ServerConfig[];
}

export type GetBotLogsParams = {
  limit?: number;
};