# lib/api-zod/src/generated/types/botTask.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Minecraft Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { BotTaskParams } from "./botTaskParams";
import type { BotTaskStatus } from "./botTaskStatus";
import type { BotTaskType } from "./botTaskType";

export interface BotTask {
  id: string;
  type: BotTaskType;
  status: BotTaskStatus;
  params: BotTaskParams;
}