# lib/api-zod/src/generated/types/commandRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Minecraft Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { CommandRequestCommand } from "./commandRequestCommand";

export interface CommandRequest {
  command: CommandRequestCommand;
  value?: string | null;
}