# lib/api-zod/src/generated/types/serverConfigsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Minecraft Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { ServerConfig } from "./serverConfig";

export interface ServerConfigsResponse {
  configs: ServerConfig[];
}