# lib/api-zod/src/generated/types/taskQueueResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Minecraft Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { BotTask } from "./botTask";

export interface TaskQueueResponse {
  tasks: BotTask[];
}